rename CKpreview to editorPreview
